// src/types/pic.d.ts
export interface CaseEvent {
  /**
   * database id
   */
  id: number;
  /**
   * Link to information about the event.
   */
  source?: string;
  /**
   * Event class (e.g., NOI, ROD)
   */
  type: string;
  /**
   * Optional event hierarchy (e.g., Tier 1 NOI, Tier 2 Scoping Complete, Tier 3 Task assigned to Jane Doe).
   */
  tier?: number;
  /**
   * Pending, completed, in progress.
   */
  status?: "pending" | "completed" | "in progress";
  /**
   * Result or action taken.
   */
  outcome?: string;
  /**
   *  Reference to Document ID of the document related to the event. Optional.
   */
  related_document_id?: number;
  /**
   * Indicates whether events can be displayed publicly. (Binary Y/N)
   */
  public_access?: boolean;
  /**
   * Description of event.
   */
  description?: string;
  /**
   * Responsible individuals or agencies.
   */
  assigned_entity?: string;
  /**
   * Name of segment following this event (e.g. if this is "start of scoping" the next segment is "scoping").
   */
  following_segment_name?: string;
  /**
   * If applicable, the parent ID of the event.
   */
  parent_event_id?: number;
  /**
   * Related engagement event if applicable
   */
  related_engagement_id?: number;
  /**
   * This is the version of this data record.
   */
  data_record_version?: string;
  /**
   * This is the agency from which this data record was sent or retrieved or in which it is stored, which may be different than the record owner.
   */
  data_source_agency?: string;
  /**
   * This is the system from which this data record was sent or retrieved or in which it is stored.
   */
  data_source_system?: string;
  /**
   * This is the last update of the data record in the system.
   */
  last_updated?: string;
  /**
   * This is the owner of the original data, generally the authoritative data source.
   */
  record_owner_agency?: string;
  /**
   * This is the timestamp of a retrieval or transmission (e.g. responding to an API call).
   */
  retrieved_timestamp?: string;
  /**
   * created
   */
  created_at?: string;
  /**
   * Reference to Process ID of the associated NEPA process for this event.
   */
  parent_process_id: number;
  /**
   * Name of event
   */
  name: string;
  /**
   * Date and time of event
   */
  datetime: string;
  /**
   * This jsonb field should be used for additional data structures as needed.
   */
  other?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}